Move :build-logic:basics to :build-logic-commons
