+++ build-logic-commons/basics/src/main/kotlin/gradlebuild.minify.gradle.kts
