+++ build-logic-commons/basics/src/main/kotlin/gradlebuild/basics/GradleModuleApiAttribute.kt
