+++ build-logic-commons/basics/src/main/kotlin/gradlebuild/basics/StringExtensions.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 the original author or authors.
+ * Copyright 2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package gradlebuild
+package gradlebuild.basics
 
 import java.util.Locale
