+++ build-logic-commons/basics/src/main/kotlin/gradlebuild/basics/classanalysis/ClassGraph.kt
