+++ build-logic-commons/basics/src/main/kotlin/gradlebuild/basics/tasks/ClasspathManifest.kt
