+++ build-logic-commons/basics/src/main/kotlin/gradlebuild/basics/util/KotlinSourceParser.kt
