+++ build-logic-commons/basics/src/main/kotlin/gradlebuild/basics/util/ReproduciblePropertiesWriter.kt
