+++ build-logic-commons/basics/src/test/kotlin/gradlebuild/basics/tasks/PackageListGeneratorIntegrationTest.kt
