+++ build-logic-commons/gradle-plugin/build.gradle.kts
@@ -18,6 +18,7 @@ dependencies {
 
     api(platform(project(":build-platform")))
 
+    implementation(project(":basics"))
     implementation(project(":commons"))
     implementation("org.gradle.kotlin.kotlin-dsl:org.gradle.kotlin.kotlin-dsl.gradle.plugin:4.1.2")
     // This Kotlin version should only be updated when updating the above kotlin-dsl version
