+++ build-logic/build-update-utils/build.gradle.kts
@@ -6,7 +6,7 @@ plugins {
 description = "Provides plugins that create update tasks for the Gradle build"
 
 dependencies {
+    implementation("gradlebuild:basics")
     implementation("com.google.code.gson:gson")
     implementation("org.jsoup:jsoup")
-    implementation(project(":basics"))
 }
