+++ build-logic/build-update-utils/src/main/kotlin/gradlebuild.wrapper.gradle.kts
@@ -1,4 +1,4 @@
-import gradlebuild.capitalize
+import gradlebuild.basics.capitalize
 import com.google.gson.Gson
 import java.net.URL
