+++ build-logic/buildquality/src/main/kotlin/gradlebuild.incubation-report-aggregation.gradle.kts
@@ -1,4 +1,4 @@
-import gradlebuild.capitalize
+import gradlebuild.basics.capitalize
 import gradlebuild.incubation.tasks.IncubatingApiAggregateReportTask
 
 plugins {
