+++ build-logic/dependency-modules/build.gradle.kts
@@ -5,6 +5,6 @@ plugins {
 description = "Provides a plugin to minify and correct metadata for dependencies used by Gradle"
 
 dependencies {
-    implementation(project(":basics"))
+    implementation("gradlebuild:basics")
     implementation("com.google.code.gson:gson")
 }
