+++ build-logic/integration-testing/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 description = "Provides plugins to create and configure integration, cross-version and distribution tests"
 
 dependencies {
-    implementation(project(":basics"))
+    implementation("gradlebuild:basics")
     implementation(project(":cleanup"))
     implementation(project(":dependency-modules"))
     implementation(project(":module-identity"))
