+++ build-logic/jvm/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 description = "Provides plugins to configure conventions used by Kotlin, Java and Groovy projects to build Gradle"
 
 dependencies {
-    implementation(project(":basics"))
+    implementation("gradlebuild:basics")
     implementation(project(":dependency-modules"))
     implementation(project(":module-identity"))
