+++ build-logic/kotlin-dsl/build.gradle.kts
@@ -5,7 +5,8 @@ plugins {
 description = "Provides plugins to configure Kotlin DSL and patch the Kotlin compiler for use in Kotlin subprojects"
 
 dependencies {
-    implementation(project(":basics"))
+    implementation("gradlebuild:basics")
+
     implementation(project(":dependency-modules"))
     implementation(project(":jvm"))
     implementation(project(":kotlin-dsl-shared-runtime"))
