+++ build-logic/lifecycle/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 description = "Provides a plugin to define entry point lifecycle tasks used for development (e.g., sanityCheck)"
 
 dependencies {
-    implementation(project(":basics"))
+    implementation("gradlebuild:basics")
 
     implementation("com.google.code.gson:gson")
 }
