+++ build-logic/packaging/build.gradle.kts
@@ -5,11 +5,12 @@ plugins {
 description = "Provides a plugin for building Gradle distributions"
 
 dependencies {
+    implementation("gradlebuild:basics")
+
     implementation(project(":documentation")) {
         // TODO turn this around: move corresponding code to this project and let docs depend on it
         because("API metadata generation is part of the DSL guide")
     }
-    implementation(project(":basics"))
     implementation(project(":module-identity"))
     implementation(project(":jvm"))
     implementation(project(":kotlin-dsl"))
