+++ build-logic/packaging/src/main/kotlin/gradlebuild.shaded-jar.gradle.kts
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
+import gradlebuild.basics.capitalize
 import gradlebuild.basics.classanalysis.Attributes
-import gradlebuild.capitalize
-import gradlebuild.decapitalize
+import gradlebuild.basics.decapitalize
 import gradlebuild.shade.ArtifactTypes.buildReceiptType
 import gradlebuild.shade.ArtifactTypes.classTreesType
 import gradlebuild.shade.ArtifactTypes.entryPointsType
