+++ build-logic/publishing/src/main/kotlin/gradlebuild.kotlin-dsl-plugin-bundle.gradle.kts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import gradlebuild.capitalize
+import gradlebuild.basics.capitalize
 import gradlebuild.pluginpublish.extension.PluginPublishExtension
 import java.time.Year
