+++ build-logic/root-build/build.gradle.kts
@@ -5,8 +5,9 @@ plugins {
 description = "Provides plugins that configures the root Gradle project"
 
 dependencies {
+    implementation("gradlebuild:basics")
+
     implementation(project(":idea"))
-    implementation(project(":basics"))
     implementation(project(":profiling"))
 
     implementation(project(":cleanup")) {
