+++ build-logic/uber-plugins/build.gradle.kts
@@ -5,7 +5,8 @@ plugins {
 description = "Provides plugins that combine and configure other plugins for different kinds of Gradle subprojects"
 
 dependencies {
-    implementation(project(":basics"))
+    implementation("gradlebuild:basics")
+
     implementation(project(":buildquality"))
     implementation(project(":cleanup"))
     implementation(project(":dependency-modules"))
